drop dependency on CLASS module
The CLASS module adds a trivial convenience of being able to use $CLASS
rather than __PACKAGE__. The CLASS module has an uncertain future,
because it currently has a maintainer attempting to combine it with a
massive complex object system, reusing the name for something entirely
unrelated. Since it is trivial to do without, just remove the
prerequisite.