remove gitpan and Git::CPAN::Patch::Import
The gitpan script does not get installed. It serves essentially no
purpose in the shipped tarball. Git::CPAN::Patch::Import exists only for
the gitpan script, so it is similarly useless. It shares no code with
the rest of the dist, so there is no reason to ship them together. It
adds several additional prereqs that serve no purpose to the installed
code. One of the prereqs, BackPAN::Index, is broken on many machines and
is unmaintained.

Preserving gitpan is best done as a separate dist.