use CPAN::DistnameInfo rather than CPAN::ParseDistribution
CPAN::DistnameInfo has less prereqs and matches the behavior of PAUSE.
It is the standard used most other places in the CPAN ecosystem for this
job.

This also fixes a bug, as CPAN::ParseDistribution was being given the
name of a temporary downloaded file, rather than the full CPAN path
needed.